+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Production: `docker-compose.prod.yml`
 - `ENV` variables for web server config 
+- Nginx: SSL certificate generation for development.
 
 ### Changed